refactor: move pushHistory into common
